📚 Renaming Matrix class

+++ Modules/Dementia/Source/Math/Frustrum.cpp
@@ -91,7 +91,7 @@ Frustum::Frustum()
 
 void Frustum::TransformFrustum(const DirectX::XMFLOAT4X4& proj, const DirectX::XMFLOAT4X4& view)
 {
-	double clip[4][4];
+	float clip[4][4];
 	clip[0][0] = (view(0, 0) * proj(0, 0) + view(0, 1) * proj(1, 0) + view(0, 2) * proj(2, 0) + view(0, 3) * proj(3, 0));
 	clip[0][1] = view(0, 0) * proj(0, 1) + view(0, 1) * proj(1, 1) + view(0, 2) * proj(2, 1) + view(0, 3) * proj(3, 1);
 	clip[0][2] = view(0, 0) * proj(0, 2) + view(0, 1) * proj(1, 2) + view(0, 2) * proj(2, 2) + view(0, 3) * proj(3, 2);
